retail_sales_analysis_mage/requirements.txt
mage[dbt]
kaggle
pyspark

